requirements.txt
asgiref==3.5.2
Django==4.0.7
djangorestframework==3.14.0
psycopg2==2.9.4
pytz==2022.6
sqlparse==0.4.3
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
urllib3==1.26.12